refactor(-): :hammer: refatora rotas GA pt2
